+++ frontend/openchat-agent/src/services/notifications/candid/types.d.ts
@@ -635,6 +635,7 @@ export interface Icrc1PendingCryptoTransaction {
   'memo' : [] | [Memo],
   'ledger' : CanisterId,
   'amount' : bigint,
+  'symbol' : string,
 }
 export interface ImageContent {
   'height' : number,
@@ -791,6 +792,7 @@ export interface NnsPendingCryptoTransaction {
   'memo' : [] | [bigint],
   'ledger' : CanisterId,
   'amount' : Tokens,
+  'symbol' : string,
 }
 export interface NnsProposal {
   'id' : ProposalId,
@@ -1033,6 +1035,7 @@ export interface SnsPendingCryptoTransaction {
   'memo' : [] | [bigint],
   'ledger' : CanisterId,
   'amount' : Tokens,
+  'symbol' : string,
 }
 export interface SnsProposal {
   'id' : ProposalId,