+++ frontend/openchat-agent/src/services/online/candid/types.d.ts
@@ -635,6 +635,7 @@ export interface Icrc1PendingCryptoTransaction {
   'memo' : [] | [Memo],
   'ledger' : CanisterId,
   'amount' : bigint,
+  'symbol' : string,
 }
 export interface ImageContent {
   'height' : number,
@@ -801,6 +802,7 @@ export interface NnsPendingCryptoTransaction {
   'memo' : [] | [bigint],
   'ledger' : CanisterId,
   'amount' : Tokens,
+  'symbol' : string,
 }
 export interface NnsProposal {
   'id' : ProposalId,
@@ -1036,6 +1038,7 @@ export interface SnsPendingCryptoTransaction {
   'memo' : [] | [bigint],
   'ledger' : CanisterId,
   'amount' : Tokens,
+  'symbol' : string,
 }
 export interface SnsProposal {
   'id' : ProposalId,