+++ backend/canisters/user/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Mark channels as read after joining community or channel ([#4004](https://github.com/open-chat-labs/open-chat/pull/4004))
 - Convert SNS transaction messages into ICRC1 messages ([#4015](https://github.com/open-chat-labs/open-chat/pull/4015))
 - Add `symbol` to pending crypto transactions ([#4018](https://github.com/open-chat-labs/open-chat/pull/4018))
+- Migrate group references to channel references after import ([#4019](https://github.com/open-chat-labs/open-chat/pull/4019))
 
 ### Removed