+++ backend/canisters/user/CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Convert SNS transaction messages into ICRC1 messages ([#4015](https://github.com/open-chat-labs/open-chat/pull/4015))
 - Add `symbol` to pending crypto transactions ([#4018](https://github.com/open-chat-labs/open-chat/pull/4018))
 - Migrate group references to channel references after import ([#4019](https://github.com/open-chat-labs/open-chat/pull/4019))
+- Trim messages before pushing them as notifications ([#4020](https://github.com/open-chat-labs/open-chat/pull/4020))
 
 ### Removed