+++ backend/canisters/user/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Return imported groups in summary updates ([#4082](https://github.com/open-chat-labs/open-chat/pull/4082))
+
 ## [[2.0.763](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.763-user)] - 2023-07-28
 
 ### Changed