+++ backend/canisters/user/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Return imported groups in summary updates ([#4082](https://github.com/open-chat-labs/open-chat/pull/4082))
+- Return all channel updates after importing to community ([#4087](https://github.com/open-chat-labs/open-chat/pull/4087))
 
 ## [[2.0.763](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.763-user)] - 2023-07-28